+++ Fpl/FplParser.Tests/TestClassInheritanceTypes.fs
@@ -6,57 +6,64 @@ open Microsoft.VisualStudio.TestTools.UnitTesting
 
 [<TestClass>]
 type TestClassInheritanceTypes () =
-
+    let replaceWhiteSpace (input: string) =
+        let whiteSpaceChars = [|' '; '\t'; '\n'; '\r'|]
+        input.Split(whiteSpaceChars)
+            |> String.concat ""
 
     [<TestMethod>]
     member this.TestSpecificType3 () =
         let result = run specificClassType """object"""
         let actual = sprintf "%O" result
-        let expected = """Success: ObjectType""".Replace("\r","")
-        Assert.AreEqual(expected, actual);
+        let expected = """Success: ObjectType"""
+        Assert.AreEqual(replaceWhiteSpace expected, replaceWhiteSpace actual);
 
     [<TestMethod>]
     member this.TestSpecificType4 () =
         let result = run specificClassType """tpl"""
         let actual = sprintf "%O" result
-        let expected = """Success: TemplateType "tpl" """.Trim().Replace("\r","")
-        Assert.AreEqual(expected, actual);
+        let expected = """Success: TemplateType "tpl" """.Trim()
+        Assert.AreEqual(replaceWhiteSpace expected, replaceWhiteSpace actual);
 
     [<TestMethod>]
     member this.TestSpecificType5 () =
         let result = run specificClassType """tplSetElem"""
         let actual = sprintf "%O" result
-        let expected = """Success: TemplateType "tplSetElem" """.Trim().Replace("\r","")
-        Assert.AreEqual(expected, actual);
+        let expected = """Success: TemplateType "tplSetElem" """.Trim()
+        Assert.AreEqual(replaceWhiteSpace expected, replaceWhiteSpace actual);
 
 
     member this.TestSpecificType7 () =
         let result = run specificClassType """@extNat"""
         let actual = sprintf "%O" result
-        let expected = """Success: ExtensionType (Extensionname "Nat")""".Replace("\r","")
-        Assert.AreEqual(expected, actual);
+        let expected = """Success: ExtensionType (Extensionname "Nat")"""
+        Assert.AreEqual(replaceWhiteSpace expected, replaceWhiteSpace actual);
 
     [<TestMethod>]
     member this.TestSpecificType8 () =
         let result = run specificClassType """SomeClass"""
         let actual = sprintf "%O" result
-        let expected = """Success: ClassHeaderType ["SomeClass"]""".Replace("\r","")
-        Assert.AreEqual(expected, actual);
+        let expected = """Success: ClassHeaderType ["SomeClass"]"""
+        Assert.AreEqual(replaceWhiteSpace expected, replaceWhiteSpace actual);
 
     [<TestMethod>]
     member this.TestSpecificType9 () =
         let result = run specificClassType """bla"""
         let actual = sprintf "%O" result
-        let actual2 = actual.Replace("\r","")
-        let expected = "Failure:\nError in Ln: 1 Col: 1\nbla\n^\nExpecting: <PascalCaseId>, @ext<PascalCaseId>, 'obj', 'object', 'template' or\n'tpl'\n"
-        Assert.AreEqual(expected, actual2);
+        let expected = "Failure:
+Error in Ln: 1 Col: 1
+bla
+^
+Expecting: <PascalCaseId>, @ext<PascalCaseId>, 'obj', 'object', 'template' or
+'tpl'"
+        Assert.AreEqual(replaceWhiteSpace expected, replaceWhiteSpace actual);
 
     [<TestMethod>]
     member this.TestClassType3 () =
         let result = run classType """object[self]"""
         let actual = sprintf "%O" result
-        let expected = """Success: FplTypeWithCoords (ObjectType, [Self []])""".Replace("\r","")
-        Assert.AreEqual(expected, actual);
+        let expected = """Success: FplTypeWithCoords (ObjectType, [Self []])"""
+        Assert.AreEqual(replaceWhiteSpace expected, replaceWhiteSpace actual);
 
     [<TestMethod>]
     member this.TestClassType3a () =
@@ -65,8 +72,8 @@ type TestClassInheritanceTypes () =
         let expected = """Success: FplTypeWithCoords
   (ObjectType,
    [AliasedId ["SomeObject1"]; AliasedId ["SomeObject2"];
-    AliasedId ["SomeObject3"]])""".Replace("\r","")
-        Assert.AreEqual(expected, actual);
+    AliasedId ["SomeObject3"]])"""
+        Assert.AreEqual(replaceWhiteSpace expected, replaceWhiteSpace actual);
 
     [<TestMethod>]
     member this.TestClassType4 () =
@@ -74,8 +81,8 @@ type TestClassInheritanceTypes () =
         let actual = sprintf "%O" result
         let expected = """Success: FplTypeWithRange
   ((TemplateType "tpl", LeftClosed),
-   (RangeInType (Some (Var "from"), None), RightClosed))""".Replace("\r","")
-        Assert.AreEqual(expected, actual);
+   (RangeInType (Some (Var "from"), None), RightClosed))"""
+        Assert.AreEqual(replaceWhiteSpace expected, replaceWhiteSpace actual);
 
     [<TestMethod>]
     member this.TestClassType4a () =
@@ -83,16 +90,16 @@ type TestClassInheritanceTypes () =
         let actual = sprintf "%O" result
         let expected = """Success: FplTypeWithRange
   ((TemplateType "tpl", LeftClosed),
-   (RangeInType (None, Some (Var "to")), RightClosed))""".Replace("\r","")
-        Assert.AreEqual(expected, actual);
+   (RangeInType (None, Some (Var "to")), RightClosed))"""
+        Assert.AreEqual(replaceWhiteSpace expected, replaceWhiteSpace actual);
 
     [<TestMethod>]
     member this.TestClassType4b () =
         let result = run classType """tpl[~]"""
         let actual = sprintf "%O" result
         let expected = """Success: FplTypeWithRange
-  ((TemplateType "tpl", LeftClosed), (RangeInType (None, None), RightClosed))""".Replace("\r","")
-        Assert.AreEqual(expected, actual);
+  ((TemplateType "tpl", LeftClosed), (RangeInType (None, None), RightClosed))"""
+        Assert.AreEqual(replaceWhiteSpace expected, replaceWhiteSpace actual);
 
     [<TestMethod>]
     member this.TestClassType5 () =
@@ -100,8 +107,8 @@ type TestClassInheritanceTypes () =
         let actual = sprintf "%O" result
         let expected = """Success: FplTypeWithRange
   ((ClassHeaderType ["Set"], LeftClosed),
-   (RangeInType (Some (Var "from"), Some (Var "to")), RightClosed))""".Replace("\r","")
-        Assert.AreEqual(expected, actual);
+   (RangeInType (Some (Var "from"), Some (Var "to")), RightClosed))"""
+        Assert.AreEqual(replaceWhiteSpace expected, replaceWhiteSpace actual);
 
     [<TestMethod>]
     member this.TestClassType5a () =
@@ -109,8 +116,8 @@ type TestClassInheritanceTypes () =
         let actual = sprintf "%O" result
         let expected = """Success: FplTypeWithRange
   ((ClassHeaderType ["Set"], LeftOpen),
-   (RangeInType (Some (Var "from"), Some (Var "to")), RightClosed))""".Replace("\r","")
-        Assert.AreEqual(expected, actual);
+   (RangeInType (Some (Var "from"), Some (Var "to")), RightClosed))"""
+        Assert.AreEqual(replaceWhiteSpace expected, replaceWhiteSpace actual);
 
     [<TestMethod>]
     member this.TestClassType5b () =
@@ -118,28 +125,32 @@ type TestClassInheritanceTypes () =
         let actual = sprintf "%O" result
         let expected = """Success: FplTypeWithRange
   ((ClassHeaderType ["Set"], LeftClosed),
-   (RangeInType (Some (Var "from"), Some (Var "to")), RightOpen))""".Replace("\r","")
-        Assert.AreEqual(expected, actual);
+   (RangeInType (Some (Var "from"), Some (Var "to")), RightOpen))"""
+        Assert.AreEqual(replaceWhiteSpace expected, replaceWhiteSpace actual);
 
 
     [<TestMethod>]
     member this.TestClassType7 () =
         let result = run classType """@extNat"""
         let actual = sprintf "%O" result
-        let expected = """Success: ExtensionType (Extensionname "Nat")""".Replace("\r","")
-        Assert.AreEqual(expected, actual);
+        let expected = """Success: ExtensionType (Extensionname "Nat")"""
+        Assert.AreEqual(replaceWhiteSpace expected, replaceWhiteSpace actual);
 
     [<TestMethod>]
     member this.TestClassType8 () =
         let result = run classType """SomeClass"""
         let actual = sprintf "%O" result
-        let expected = """Success: ClassHeaderType ["SomeClass"]""".Replace("\r","")
-        Assert.AreEqual(expected, actual);
+        let expected = """Success: ClassHeaderType ["SomeClass"]"""
+        Assert.AreEqual(replaceWhiteSpace expected, replaceWhiteSpace actual);
 
     [<TestMethod>]
     member this.TestClassType9 () =
         let result = run classType """bla"""
         let actual = sprintf "%O" result
-        let actual2 = actual.Replace("\r","")
-        let expected = "Failure:\nError in Ln: 1 Col: 1\nbla\n^\nExpecting: <PascalCaseId>, @ext<PascalCaseId>, 'obj', 'object', 'template' or\n'tpl'\n"
-        Assert.AreEqual(expected, actual2);
+        let expected = "Failure:
+Error in Ln: 1 Col: 1
+bla
+^
+Expecting: <PascalCaseId>, @ext<PascalCaseId>, 'obj', 'object', 'template' or
+'tpl'"
+        Assert.AreEqual(replaceWhiteSpace expected, replaceWhiteSpace actual);